Movement of the larger test sources into separate resource files for easier editing
